requirements.txt
pywinauto
pytesseract
Pillow
opencv-python
pyautogui
pyperclip
keyboard
